+++ src/ValidationServiceProvider.php
@@ -2,12 +2,26 @@
 
 namespace Axlon\PostalCodeValidation;
 
-use Axlon\PostalCodeValidation\Extensions\PostalCodeFor;
+use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Validation\Factory;
 use Illuminate\Support\ServiceProvider;
 
 class ValidationServiceProvider extends ServiceProvider
 {
+    /**
+     * Boot postal code validation services.
+     *
+     * @return void
+     */
+    public function boot(): void
+    {
+        $this->mergeConfigFrom($configPath = __DIR__ . '/../config/postal_codes.php', 'postal_codes');
+
+        if ($this->app->runningInConsole()) {
+            $this->publishes([$configPath => config_path('postal_codes.php')], 'config');
+        }
+    }
+
     /**
      * Register postal code validation services.
      *
@@ -23,8 +37,13 @@ class ValidationServiceProvider extends ServiceProvider
             });
         }
 
-        $this->app->singleton(PostalCodeFor::class);
-        $this->app->singleton(Validator::class);
+        $this->app->singleton('postal_codes', function (Application $app) {
+            return new PatternMatcher(
+                require __DIR__ . '/../resources/patterns.php', $app['config']['postal_codes.overrides']
+            );
+        });
+
+        $this->app->alias('postal_codes', PatternMatcher::class);
     }
 
     /**