+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [v3.0.0](https://github.com/axlon/laravel-postal-code-validation/tree/v3.0.0) (TBD)
 - Dropped support for PHP 7.1
 - Dropped support for Laravel 5.1 - 5.4
+- Added support for manually overriding validation patterns
+- Added lazy loading for postal code examples
+- Added exceptions when no arguments are passed to the validation rules
 
 ## [v2.1.0](https://github.com/axlon/laravel-postal-code-validation/tree/v2.1.0) (2020-03-04)
 - Added support for Laravel 7